+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 0.0.8 (SNAPSHOT)
 * Add support for creating configuration file with GitHub repositories
 * Add undo command
+* Add clean command
 
 ### 0.0.7
 * Add custom config file override